renames package analyzer => analyzers
