+++ src/dimcat/steps/analyzers/__init__.py
@@ -1,0 +1,6 @@
+import logging
+
+from .counters import Counter
+from .pitch import PitchClassVectors
+
+logger = logging.getLogger(__name__)
