+++ src/dimcat/steps/analyzers/base.py
