+++ src/dimcat/steps/analyzers/pitch.py
@@ -4,7 +4,7 @@ from dimcat.data.resource.base import SomeDataframe, SomeSeries
 from dimcat.data.resource.dc import DimcatResource
 from dimcat.data.resource.features import Feature, FeatureName
 from dimcat.data.resource.results import Durations
-from dimcat.steps.analyzer.base import Analyzer
+from dimcat.steps.analyzers.base import Analyzer
 
 logger = logging.getLogger(__name__)
